app: move Client to app.internal.client
